pyproject.toml
[project]
name = "give-me-money"
version = "0.1.0"
description = "My experimental/learning project for options trading"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "args==0.1.0",
    "attrs==25.4.0",
    "beautifulsoup4==4.14.3",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "clint==0.5.1",
    "colorama==0.4.6",
    "frozenlist==1.8.0",
    "google==3.0.0",
    "growwapi==1.5.0",
    "idna==3.11",
    "iniconfig==2.3.0",
    "multidict==6.7.1",
    "nats-py==2.13.1",
    "nkeys==0.2.1",
    "numpy==2.4.2",
    "packaging==26.0",
    "pandas==2.3.3",
    "pluggy==1.6.0",
    "propcache==0.4.1",
    "protobuf==5.29.6",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "pynacl==1.6.2",
    "pyotp==2.9.0",
    "pytest==9.0.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.2.1",
    "pytz==2025.2",
    "requests==2.32.5",
    "six==1.17.0",
    "soupsieve==2.8.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "yarl==1.22.0",
]
